hubble: Add Exynos7580 Support, and support for detecting legacy SoCs and fix cleanup for legacy SoCs
This change adds Support for the Exynos7580 and adds legacy SoC detection.

Legacy SoCs report back the Exynos4210 SoC name in their device descriptor,
this is incorrect for SoCs like the Exynos7580, which clearly is not the
Exynos4210. Instead, when we detect an Exynos4210 chipset, we flag the
device as legacy and run an alternative detection routine that parses the
sboot.bin file with to hopefully find the right SoC.

File cleanup has to be fixed too as legacy SoCs don't have lz4 compressed
images, just raw bins, which hubble wasn't able to handle, this commit
adds handling for cleanup to work on the data of legacy SoCs.

Tested-by: Daniel <danalexgro@gmail.com>
Tested-by: Umer Uddin <umer.uddin@mentallysanemainliners.org>

[Added Exynos7580 support]
Co-authored-by: Robotix <98209628+Robotix22@users.noreply.github.com>

[Added Legacy SoC detection and fixed cleanup]
Signed-off-by: Umer Uddin <umer.uddin@mentallysanemainliners.org>